+++ CHANGELOG.md
@@ -5,3 +5,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- ability to load kml and csv (unicsv format) files
+  [#1](https://gitlab.com/eneiluj/gpxedit-oc/issues/1) @eneiluj